requirements.txt
tensorboard
matplotlib==3.4.3
Pillow==9.1.1
torch==1.11.0
torchinfo==1.7.0
torchvision==0.12.0+cu113
tqdm==4.62.3
~orch==1.8.0
